requirements.txt
absl-py==0.13.0
aio-pika==6.8.0
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.4
boto3==1.20.14
botocore==1.23.14
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.8
clang==5.0
cloudpickle==1.6.0
colorclass==2.2.0
coloredlogs==15.0.1
colorhash==1.0.3
cryptography==36.0.0
cycler==0.11.0
dask==2021.7.2
decorator==5.1.0
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.4.0
flatbuffers==1.12
fsspec==2021.11.1
future==0.18.2
gast==0.5.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h5py==3.1.0
httptools==0.3.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.8.2
jmespath==0.10.0
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
locket==0.2.1
Markdown==3.3.6
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.3
multidict==5.2.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
partd==1.2.0
Pillow==8.4.0
prompt-toolkit==2.0.10
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyJWT==2.3.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.6
pyrsistent==0.18.0
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.3
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.0.0
rasa-sdk==3.0.0
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rocketchat-API==1.16.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sanic==21.9.3
Sanic-Cors==1.0.1
sanic-jwt==1.7.0
sanic-plugin-toolkit==1.2.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
sentry-sdk==1.3.1
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
SQLAlchemy==1.4.27
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-probability==0.13.0
tensorflow-text==2.6.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==3.0.0
toolz==0.11.2
tqdm==4.62.3
twilio==6.50.1
typeguard==2.13.2
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.7
uvloop==0.14.0
wcwidth==0.2.5
webexteamssdk==1.6
websockets==10.1
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0
